requirements.txt
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
albumentations>=1.3.0
hydra-core>=1.3.0
wandb>=0.15.0
rich>=13.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
omegaconf>=2.3.0
torchmetrics>=1.0.0
onnx>=1.14.0
onnxruntime>=1.16.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
pandas>=2.0.0
fastparquet>=2023.11.0